+++ client/tests/components/GameCard.test.tsx
@@ -2,7 +2,7 @@ import { describe, it, expect } from "vitest";
 import { render, screen, within } from "@testing-library/react";
 //import { userEvent } from "@testing-library/user-event";
 import React from "react";
-import GameCard from "../../src/components/gameCard/GameCard";
+import GameCard from "../../src/components/GameCard/GameCard";
 import "@testing-library/jest-dom";
 import { socket } from "../../src/services/socketServices";
